requirements.txt
asgiref==3.8.1
Django==5.0.6
django-admin-interface==0.28.6
django-cleanup==8.1.0
django-colorfield==0.11.0
django-filter==24.2
django-resized==1.0.2
djangorestframework==3.15.1
Markdown==3.6
pillow==10.3.0
python-slugify==8.0.4
sqlparse==0.5.0
text-unidecode==1.3
typing_extensions==4.12.0
